+++ packages/agent/src/agent.ts
@@ -19,6 +19,7 @@ import bodyParser from 'koa-bodyparser';
 import FrameworkMounter from './framework-mounter';
 import makeRoutes from './routes';
 import makeServices from './services';
+import ActionCustomizationService from './services/model-customizations/action-customization';
 import { AgentOptions, AgentOptionsWithDefaults } from './types';
 import SchemaGenerator from './utils/forest-schema/generator';
 import OptionsValidator from './utils/options-validator';
@@ -36,6 +37,8 @@ import OptionsValidator from './utils/options-validator';
 export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter {
   private options: AgentOptionsWithDefaults;
   private customizer: DataSourceCustomizer<S>;
+  private nocodeCustomizer: DataSourceCustomizer<S>;
+  private actionCustomizationService: ActionCustomizationService;
 
   /**
    * Create a new Agent Builder.
@@ -59,6 +62,9 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
 
     this.options = allOptions;
     this.customizer = new DataSourceCustomizer<S>();
+    this.nocodeCustomizer = new DataSourceCustomizer<S>();
+    this.nocodeCustomizer.addDataSource(this.customizer.getFactory());
+    this.actionCustomizationService = new ActionCustomizationService(allOptions);
   }
 
   /**
@@ -160,9 +166,15 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
   }
 
   private async getRouterAndSendSchema(): Promise<Router> {
-    const { isProduction, logger, skipSchemaUpdate, typingsPath, typingsMaxDepth } = this.options;
+    const { isProduction, logger, skipSchemaUpdate, typingsPath, typingsMaxDepth, experimental } =
+      this.options;
 
-    const dataSource = await this.customizer.getDataSource(logger);
+    await this.nocodeCustomizer.use(
+      this.actionCustomizationService.addWebhookActions,
+      experimental?.webhookCustomActions,
+    );
+
+    const dataSource = await this.nocodeCustomizer.getDataSource(logger);
     const [router] = await Promise.all([
       this.getRouter(dataSource),
       // skipSchemaUpdate is mainly used in cloud version
@@ -191,7 +203,7 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
         throw new Error(`Can't load ${schemaPath}. Providing a schema is mandatory in production.`);
       }
     } else {
-      schema = await SchemaGenerator.buildSchema(dataSource);
+      schema = await SchemaGenerator.buildSchema(dataSource, this.buildSchemaFeatures());
 
       const pretty = stringify(schema, { maxLength: 100 });
       await writeFile(schemaPath, pretty, { encoding: 'utf-8' });
@@ -204,5 +216,15 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
       : 'Schema was not updated since last run';
 
     this.options.logger('Info', message);
+  }
+
+  private buildSchemaFeatures(): string[] | null {
+    const mapping: Record<keyof AgentOptions['experimental'], string> = {
+      webhookCustomActions: ActionCustomizationService.FEATURE,
+    };
+
+    return Object.entries(mapping)
+      .filter(([experimentalFeature]) => this.options.experimental?.[experimentalFeature])
+      .map(([, feature]) => feature);
   }
 }