+++ packages/forestadmin-client/test/__factories__/forest-admin-client.ts
@@ -1,9 +1,13 @@
 import { Factory } from 'fishery';
 
+import authServiceFactory from './auth';
 import chartHandlerFactory from './chart/chart-handler';
 import forestAdminClientOptionsFactory from './forest-admin-client-options';
+import ipWhitelistServiceFactory from './ip-whitelist';
+import modelCustomizationServiceFactory from './model-customizations/model-customization-from-api';
 import permissionServiceFactory from './permissions/permission';
 import renderingPermissionsFactory from './permissions/rendering-permission';
+import schemaServiceFactory from './schema';
 import contextVariablesInstantiatorFactory from './utils/context-variables-instantiator';
 import ForestAdminClient from '../../src/forest-admin-client-with-cache';
 
@@ -25,6 +29,10 @@ const forestAdminClientFactory = ForestAdminClientFactory.define(
       renderingPermissionsFactory.build(),
       contextVariablesInstantiatorFactory.build(),
       chartHandlerFactory.build(),
+      ipWhitelistServiceFactory.build(),
+      schemaServiceFactory.build(),
+      authServiceFactory.build(),
+      modelCustomizationServiceFactory.build(),
     ),
 );