+++ packages/forestadmin-client/test/forest-admin-client-with-cache.test.ts
@@ -25,6 +25,7 @@ describe('ForestAdminClientWithCache', () => {
         whiteListService,
         factories.schema.build(),
         factories.auth.build(),
+        factories.modelCustomization.build(),
       );
 
       const config = await forestAdminClient.getIpWhitelistConfiguration();
@@ -48,6 +49,7 @@ describe('ForestAdminClientWithCache', () => {
         factories.ipWhiteList.build(),
         schemaService,
         factories.auth.build(),
+        factories.modelCustomization.build(),
       );
 
       const result = await forestAdminClient.postSchema({
@@ -55,6 +57,7 @@ describe('ForestAdminClientWithCache', () => {
         metadata: {
           liana: 'forest-nodejs-agent',
           liana_version: '1.0.0',
+          liana_features: null,
           stack: { engine: 'nodejs', engine_version: '16.0.0' },
         },
       });
@@ -76,6 +79,7 @@ describe('ForestAdminClientWithCache', () => {
         factories.ipWhiteList.build(),
         factories.schema.build(),
         factories.auth.build(),
+        factories.modelCustomization.build(),
       );
 
       verifyAndExtractApprovalMock.mockReturnValue(signedParameters);
@@ -102,6 +106,7 @@ describe('ForestAdminClientWithCache', () => {
         factories.ipWhiteList.build(),
         factories.schema.build(),
         authService,
+        factories.modelCustomization.build(),
       );
 
       const result = await forestAdminClient.getUserInfo(1, 'token');
@@ -124,6 +129,7 @@ describe('ForestAdminClientWithCache', () => {
         factories.ipWhiteList.build(),
         factories.schema.build(),
         authService,
+        factories.modelCustomization.build(),
       );
 
       const result = await forestAdminClient.getOpenIdClient();
@@ -145,6 +151,7 @@ describe('ForestAdminClientWithCache', () => {
         factories.ipWhiteList.build(),
         factories.schema.build(),
         factories.auth.build(),
+        factories.modelCustomization.build(),
       );
 
       await forestAdminClient.markScopesAsUpdated(42);
@@ -165,6 +172,7 @@ describe('ForestAdminClientWithCache', () => {
         factories.ipWhiteList.build(),
         factories.schema.build(),
         factories.auth.build(),
+        factories.modelCustomization.build(),
       );
 
       (renderingPermissionService.getScope as jest.Mock).mockResolvedValue('scope');